requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.2.1
davey==0.1.5
discord.py==2.7.1
Flask==3.1.1
frozenlist==1.7.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
mutagen==1.47.0
propcache==0.3.2
pycparser==2.22
pycryptodomex==3.23.0
PyNaCl==1.5.0
python-dotenv==1.1.1
requests==2.33.1
typing_extensions==4.14.1
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.3
yarl==1.20.1
youtube-dl==2021.12.17
yt-dlp==2026.3.17
yt-dlp-ejs==0.8.0
